# Conflicts:
#	client/lib/features/auth/presentation/widgets/sign_in_options_content.dart
#	client/lib/l10n/app_en.arb
#	client/lib/l10n/app_es.arb
#	client/lib/l10n/app_zh.arb
#	client/lib/l10n/app_zh_Hant_TW.arb

+++ client/lib/features/events/features/event_page/presentation/widgets/event_page_meeting_agenda.dart
@@ -47,7 +47,7 @@ class _EventPageMeetingAgendaState extends State<EventPageMeetingAgenda>
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
             HeightConstrainedText(
-              'Add breakouts',
+              context.l10n.addBreakouts,
               style: context.theme.textTheme.titleLarge!
                   .copyWith(color: context.theme.colorScheme.onSurfaceVariant),
             ),
@@ -191,7 +191,7 @@ class _EventPageMeetingAgendaState extends State<EventPageMeetingAgenda>
 
   Widget _buildAgendaTitle() {
     return HeightConstrainedText(
-      'Agenda',
+      context.l10n.agendaTitle,
       style: context.theme.textTheme.titleLarge,
     );
   }
@@ -200,7 +200,7 @@ class _EventPageMeetingAgendaState extends State<EventPageMeetingAgenda>
     return ActionButton(
       type: ActionButtonType.text,
       onPressed: () => _showClearAgendaItemsDialog(),
-      text: 'Clear all',
+      text: context.l10n.clearAll,
       icon: Padding(
         padding: const EdgeInsets.only(left: 5),
         child: Icon(
@@ -216,14 +216,13 @@ class _EventPageMeetingAgendaState extends State<EventPageMeetingAgenda>
   Future<void> _showClearAgendaItemsDialog() async {
     final delete = await ConfirmDialog(
       title: context.l10n.clearAgenda,
-      mainText:
-          'Are you sure you want to remove all agenda items from the breakout rooms? You won\'t be able to undo this.',
+      mainText: context.l10n.confirmRemoveAllAgendaItems,
       cancelText: context.l10n.cancel,
-      confirmText: 'Yes, clear',
+      confirmText: context.l10n.yesClear,
       textAlign: TextAlign.start,
     ).show(context: context);
 
-    if (delete) {
+    if (delete && mounted) {
       await alertOnError(context, () => _presenter.deleteAgendaItems());
     }
   }
